debugger: move ProbeInspectorSession and helpers to separate files
